requirements.txt
Django
psycopg2-binary
djangorestframework
gunicorn
python-decouple